Moved items folder to data folder
