+++ src/main/java/com/honiism/discord/lemi/data/database/LemiDbBalDs.java
@@ -30,8 +30,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.honiism.discord.lemi.Lemi;
-import com.honiism.discord.lemi.commands.slash.currency.objects.items.Items;
 import com.honiism.discord.lemi.data.database.managers.LemiDbBalManager;
+import com.honiism.discord.lemi.data.items.Items;
 import com.honiism.discord.lemi.utils.currency.CurrencyTools;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
