+++ src/main/java/com/honiism/discord/lemi/data/items/handler/EventType.java
@@ -17,13 +17,14 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.commands.slash.currency.objects.items.handler;
+package com.honiism.discord.lemi.data.items.handler;
 
-public enum ItemType {
-    SELLABLE,
-    ITEM_PACK,
-    POWER_UP,
-    COLLECTABLE,
-    TOOL,
-    LOOT_CHESTS
+public enum EventType {
+    CHINESE_NEW_YEAR,
+    VALENTNE,
+    EASTER,
+    HALLOWEEN,
+    CHRISTMAS,
+    NEW_YEAR,
+    NONE
 }
