+++ src/main/java/com/honiism/discord/lemi/data/items/handler/ItemCategory.java
@@ -17,7 +17,7 @@
  * along with Lemi-Bot. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.honiism.discord.lemi.commands.slash.currency.objects.items.handler;
+package com.honiism.discord.lemi.data.items.handler;
 
 public enum ItemCategory {
     COMMON,
